TweetFCT_V0.1/requirements.txt
tweepy==3.7.0
factom-api==0.1.1
ipython
dataset
textblob
Django>=2.0,<3.0
psycopg2>=2.7,<3.0